refactor: Move the noise source into DynamicClipping.kt
Only DynamicClipping uses it and it already built into the same target,
so the separate file bought nothing. Matches ResultNoiser.kt and
Noiser.kt, which each hold an interface alongside its implementations.

Issue: #4341
